Vulkan (and dasher) is back but with some issues
-Resizing breaks the application
-Draws are missing (possibly wrong offsets)
-Validation errors when entering no editor mode
-Considerably worse performance than Dx12

+++ src/Core/DbLog/blitLogger.h
@@ -44,7 +44,7 @@ namespace BlitzenCore
     {
 #if defined(NDEBUG)
 
-        if (!BlitzenCore::BLIT_CHECK_FAIL(level))
+        if (!BlitzenCore::BLIT_CHECK_FAIL((int64_t)level))
         {
             return;
         }
